refactor: separated interfaces of repository and service of employees into a different file
