+++ internal/repository/employee.go
@@ -28,16 +28,6 @@ const (
 		DELETE FROM employees WHERE id=?`
 )
 
-type EmployeeRepository interface {
-	Create(ctx context.Context, e *models.Employee) (*models.Employee, error)
-	FindByCardNumberID(ctx context.Context, cardNumberID string) (*models.Employee, error)
-	FindAll(ctx context.Context) ([]*models.Employee, error)
-	FindByID(ctx context.Context, id int) (*models.Employee, error)
-	Update(ctx context.Context, id int, e *models.Employee) (*models.Employee, error)
-	Delete(ctx context.Context, id int) error
-	ExistsByCardNumberID(ctx context.Context, cardNumberID string) (bool, error)
-}
-
 type EmployeeMySQLRepository struct {
 	db *sql.DB
 }
